+++ ImperatorToCK3/Source/Imperator/Provinces/Provinces.h
@@ -20,7 +20,7 @@ namespace Imperator
 	  private:
 		void registerKeys();
 
-		std::map<int, std::shared_ptr<Province>> provinces;
+		std::map<unsigned long long, std::shared_ptr<Province>> provinces;
 	};
 } // namespace Imperator
